pyproject.toml
[project]
name = "hackon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "torch",
    "torchvision",
    "ultralytics",
    "opencv-contrib-python",
    "tqdm",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "matplotlib>=3.10.8"
]